get rid of some complier warnings

+++ src/WinMgr.h
@@ -128,7 +128,7 @@
  * this where WIN_WORKSPACE_COUNT comes into play.
  */
 
-#define WinWorkspaceInvalid    ((unsigned long)-1)
+#define WinWorkspaceInvalid    -1L
 
 /* workspaces */
 #define XA_WIN_WORKSPACES "_WIN_WORKSPACES"
